news/requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
django-crispy-forms==1.11.1
django-filer==2.0.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
easy-thumbnails==2.7.1
ipdb==0.13.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.17
ptyprocess==0.7.0
Pygments==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
Unidecode==1.1.2
wcwidth==0.2.5
